server/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
distro==1.9.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SSE==1.0.0
gevent==23.9.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonify==0.5
MarkupSafe==2.1.5
numpy
pandas
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
pytz==2024.1
redis==5.2.0
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
Werkzeug==2.0.3
openai==1.43.0